feat: Make Snaps UI footers properly sticky (#13736)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This PR fixes a couple of issues with existing components, such as the
input component not rendering error text correctly. It also makes
footers properly sticky and lets the contents of Snaps UI scroll.
Furthermore, it updates the template renderer to match the allowed
inputs from the extension.

Note that once again the majority of the diff is snapshots and tests 😄 

## **Manual testing steps**

N/A

## **Screenshots/Recordings**
![Simulator Screenshot - iPhone 16 Pro - 2025-02-27 at 14 40
59](https://github.com/user-attachments/assets/b2853c6e-ba7a-40b9-8a76-216b3bc04dc9)