fix: Use proper implementation for `getUnlockPromise` (#13947)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

`getUnlockPromise` was erroneously implemented as being tied to the
React Native AppState. It should be tied to the KeyringController state.
This PR fixes that.

## **Related issues**

Closes https://github.com/MetaMask/snaps/issues/3212