feat: Implement Segment sampling for expected errors (#13874)
## **Description**

The purpose of this task is to reduce the instances of events being
logged to Segment to only 1% of the current volume. The solution is to
randomly select 1% of USERS (NOT EVENTS) to be tracked, and ignore the
other 99%.

## **Related issues**

Fixes:
[#4119](https://github.com/MetaMask/MetaMask-planning/issues/4119)

## **Manual testing steps**

To manually test you can opt-in to being tracked via the account
creation user flow, then check MixPanel to see if the events were logged
or not.

## **Screenshots/Recordings**

### **Before**


https://github.com/user-attachments/assets/63bfb33e-f2a1-4656-a60c-be1e633beff2

### **After**


https://github.com/user-attachments/assets/9c9d0f15-7927-429a-8143-65ec40779936

## **Pre-merge author checklist**

- [X] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [X] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [X] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [X] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Daniel Cross <dan.s.cross@icloud.com>