fix: toggle to display the correct balance (#13916)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

It's been reported that setting the primary currency to "Native" was
displaying fiat as the main currency and the native currency as
secondary. The opposite would happen when setting the primary currency
to "Fiat".

This was due to using the wrong nomenclature as the properties of the
component responsible for that behaviour (`balance` was the primary
balance and `mainBalance` was the secondary balance).

The nomenclature has been fixed so that `balance` remains as the primary
balance and `mainBalance` is renamed to `secondaryBalance` so that it
matches its behaviour.

## **Related issues**

Fixes: No open GitHub issue for the ticket

## **Manual testing steps**

1. Go to Settings -> General -> Primary Currency -> Set to "Native"
2. Got to Wallet Home page and check assets in the Tokens tab display
the native currency as primary and fiat as secondary.
3. Click on any of the assets to open the asset overview and check that
the balance is displayed correctly.
4. Repeat steps 1, 2 and 3 changing the primary currency to "Fiat"

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->


![image](https://github.com/user-attachments/assets/313299ad-2a55-44f1-9f0e-8be36c8a1c05)

![image](https://github.com/user-attachments/assets/e44cdbcb-84ad-4f6f-b656-b51c61190895)

### **After**

<!-- [screenshots/recordings] -->

![Screenshot_1739377743](https://github.com/user-attachments/assets/1eadd02e-1e07-4752-92cc-43b4198ab9b6)

![Screenshot_1739377733](https://github.com/user-attachments/assets/bdf00254-fb81-4b98-96b6-aaf9aed4e262)

## **Pre-merge author checklist**

- [X] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [X] I've completed the PR template to the best of my ability
- [X] I’ve included tests if applicable
- [X] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [X] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.