feat: add metrics for bridge/swap button_clicked and page_viewed (#14777)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
Adds `PageViewed` and `ButtonClicked` events for both bridge and swaps.

To note: we currently have a bit of a patchwork of bridge and swap
interfaces based on network. For EVM networks, we have the new bridge UI
and the legacy swaps UI. For Solana, both swaps and bridge use the new
bridge UI, with swaps simply being the bridge UI with a different header
and button text. Because of this, we leave the existing metrics alone on
the legacy swaps UI and use these new metrics ONLY for the new bridge
ui. The exact event fired is dependent on the state of that UI when the
event is triggered. E.g. "Swaps page viewed" if the UI is loaded on
Solana with the swaps header, "Bridge page viewed" if the UI is loaded
on any network with the bridge header.
<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this Bridge on EVM, or either swaps or bridge on solana
2. See events firing in Segment

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Ziad Saab <ziad.saab@gmail.com>